matcher/requirements.txt
psycopg2-binary==2.9.7
python-dotenv==1.0.0
schedule==1.2.0
sqlalchemy==2.0.21